GenBCode and BCodeOpt: make all numeric coercions explicit
Following the example of a2e8f9e001356a993da4e01235c12583d97163d3

+++ src/compiler/scala/tools/nsc/backend/jvm/BCodeBodyBuilder.scala
@@ -443,8 +443,8 @@ abstract class BCodeBodyBuilder extends BCodeLateClosuBuilder {
 
         case BooleanTag => bc.boolconst(const.booleanValue)
 
-        case ByteTag    => bc.iconst(const.byteValue)
-        case ShortTag   => bc.iconst(const.shortValue)
+        case ByteTag    => bc.iconst(const.byteValue.toInt)
+        case ShortTag   => bc.iconst(const.shortValue.toInt)
         case CharTag    => bc.iconst(const.charValue)
         case IntTag     => bc.iconst(const.intValue)
